refactor(secrets): extract port type-guards to *.guard.ts files (unblocks vitest 4)
Pre-commit hook ports shipped a runtime helper (the
isPreCommitHook{Install,Uninstall}Status type guard) inside the
.port.ts file alongside the interface. To keep coverage measurement
on that helper we relied on two `!`-prefixed patterns in
`vitest.config.ts > coverage.exclude` that overrode the blanket
`src/**/*.port.ts` exclusion for these two specific files.

Vitest 4 has a known bug (vitest#10164, AriPerkkio confirmed on
2026-05-07) where negation patterns inside `coverage.exclude`
poison the inclusion logic of `BaseCoverageProvider.isIncluded()`.
The result under vitest 4 is an empty `lcov.info` (0 source files,
0 bytes), which makes SonarQube's quality gate report 0% coverage
on new code and FAIL with `exit code 3` — exactly the failure we
observed on PR #50 (vitest bump 3.2.4 → 4.1.6).

This change extracts the runtime helpers (constant + type + guard
function) out of the port files into new sibling files
`*-status.guard.ts`. The `.port.ts` files now stay 100% type-only
(D-021 convention: ports are pure interface, erased by tsc).
The two `!` negations are removed from `vitest.config.ts`.

Affected files:
- NEW src/modules/secrets/application/ports/out/pre-commit-hook-installer-status.guard.ts
- NEW src/modules/secrets/application/ports/out/pre-commit-hook-uninstaller-status.guard.ts
- src/modules/secrets/application/ports/out/pre-commit-hook-installer.port.ts (now type-only)
- src/modules/secrets/application/ports/out/pre-commit-hook-uninstaller.port.ts (now type-only)
- src/modules/secrets/application/ports/index.ts (barrel re-exports guard function from new path)
- tests/unit/secrets/application/pre-commit-hook-installer-port.test.ts (import path updated)
- tests/unit/secrets/application/pre-commit-hook-uninstaller-port.test.ts (import path updated)
- vitest.config.ts (drops the two `!` negations + adds note explaining the convention)

Validation:
- vitest 3.2.4 (develop baseline): 2588 passing in 212 files. lcov.info: 483KB / 429 SF entries.
- vitest 4.1.6 + this fix (validated in this branch, then reverted): lcov.info: 279KB / 429 SF entries (matches v3 file count, including both new .guard.ts files). Before this fix under v4: 0KB / 0 SF entries.
- typecheck EXIT=0, lint EXIT=0, lint:tests EXIT=0, validate:modules EXIT=0, build EXIT=0.

Once this PR merges to develop, Dependabot's PR #50 (vitest 4 bump) will
auto-rebase and CI's SonarQube quality gate should PASS, unblocking the
vitest 4 upgrade.
